% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{as.data.frame.prms}
\alias{as.data.frame.prms}
\title{deprecated method}
\usage{
\method{as.data.frame}{prms}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object produced by function \code{fit_enorm}}

\item{row.names}{NULL or a character vector giving the row names for the data frame. 
Missing values are not allowed.}

\item{optional}{Will be set to FALSE, so ignore.}

\item{...}{any other parameters to the as.data.frame method are ignored silently}
}
\description{
deprecated, please use \code{coef()}
}
