% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{coef.inter}
\alias{coef.inter}
\title{Extract interaction model parameters}
\usage{
\method{coef}{inter}(object, what = c("items", "scoreprob"), ...)
}
\arguments{
\item{object}{an object returend by the function \code{\link{fit_inter}}}

\item{what}{which coefficients to return. Defaults to \code{items} (the item parameters), can also be \code{scoreprob}
for the probability of each item score per booklet score.}

\item{...}{further arguments to coef are ignored}
}
\description{
Extract interaction model parameters
}
