% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_cor.R
\name{coef.latent_cor}
\alias{coef.latent_cor}
\title{latent correlations}
\usage{
\method{coef}{latent_cor}(object, hpd = 0.95, ...)
}
\arguments{
\item{object}{resulting value from latent_cor}

\item{hpd}{width of the confidence envelope of the  Bayesian highest posterior density interval around the correlations, 
value must be between 0 and 1.}

\item{...}{ignored}
}
\value{
list with matrices cor, sd, hpd_l (lower boundary), hpd_u (upper boundary), array mcmc with the complete mcmc sample 

Extract correlations with a specified highest posterior density interval
}
\description{
latent correlations
}
