% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.R
\name{coef.p2pass}
\alias{coef.p2pass}
\title{extract equating information}
\usage{
\method{coef}{p2pass}(object, ...)
}
\arguments{
\item{object}{an p2pass object, generated by \code{\link{probability_to_pass}}}

\item{...}{further arguments are currently ignored}
}
\value{
A data.frame with columns:
\describe{
\item{booklet_id}{id of the target booklet}
\item{score_new}{score on the target booklet}
\item{probability_to_pass}{probability to pass on the reference test given score_new}
\item{true_positive}{proportion that correctly passes}
\item{sensitivity}{The proportion of positives that are correctly identified as such}
\item{specificity}{The proportion of negatives that are correctly identified as such}
\item{proportion}{proportion in sample with score_new}}
}
\description{
extract equating information
}
