% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{coef.rim}
\alias{coef.rim}
\title{Coefficients for the interaction model}
\usage{
\method{coef}{rim}(object, ...)
}
\arguments{
\item{object}{An object produced by function \code{fit_inter}}

\item{...}{Further arguments currently ignored}
}
\description{
Extract parameters for the interaction and the Rasch model
}
