% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{design_as_network}
\alias{design_as_network}
\alias{design_is_connected}
\title{Deprecated test design functions}
\usage{
design_as_network(dataSrc, predicate = NULL, weights = c("items",
  "responses"))

design_is_connected(design)
}
\arguments{
\item{dataSrc}{Data source: a dexter project db handle or a data.frame with columns: person_id, item_id, item_score}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{weights}{obsolete}

\item{design}{obsolete}
}
\description{
These functions have been replaced by \code{\link{design_info}}
}
