% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{design_is_connected}
\alias{design_is_connected}
\title{Test if design is connected}
\usage{
design_is_connected(design)
}
\arguments{
\item{design}{Output from design_as_network}
}
\value{
TRUE or FALSE
}
\description{
Use the output from design_as_network to check if your design is connected.
}
\examples{
\donttest{
\dontrun{
# as an example, turn off some your booklets and see if you are
# still left with a connected design

dsgn = design_as_network(db, !(booklet_id \%in\% c('b1','b3','b4')))

design_is_connected(dsgn)
}
}

}
