% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{package}
\name{dexter}
\alias{dexter}
\alias{dexter-package}
\title{dexter: A package for doing initial data exploration on new datasets.}
\description{
The dexter package provides data exploration functions for univariate and bivariate analysis.
}
\section{univariate functions}{

All of the data exploration for univariate analysis starts with \code{\link{univariate_analysis}}.
Given a dataset it will provide useful summary statistics and plots.
Note that it does not actually render plots, but \code{\link{plot_univariates}} provides utilities to do so.
}

\section{bivariate functions}{

These function do not exist yet
}

