% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{distractor_plot}
\alias{distractor_plot}
\title{Distractor plot}
\usage{
distractor_plot(db, item, nc = 1, nr = 1)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}

\item{item}{The ID of the item to plot. A separate plot will be produced
for each booklet that contains the item, or an error message if the item ID
is not known. Each plot contains a non-parametric regression of each possible
response on the total score.}

\item{nc}{An integer between 1 and 3. Number of columns when putting mutiple plots
on the same page. Default is 1. May be ignored or adjusted if it does not make sense.}

\item{nr}{An integer between 1 and 3. Number of rows when putting mutiple plots
on the same page. Default is 1. May be ignored or adjusted if it does not make sense.}
}
\description{
Produce a diagnostic distractor plor for an item
}
\author{
Ivailo Partchev
}

