% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{fit_enorm}
\alias{fit_enorm}
\title{Fit the extended nominal response model}
\usage{
fit_enorm(dataSrc, predicate = NULL, method = c("CML", "Bayes"),
  nIterations = 500)
}
\arguments{
\item{dataSrc}{Data source: a dexter project db handle or a data.frame}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{method}{If CML, the estimation method will be Conditional Maximum Likelihood;
otherwise, a Gibbs sampler will be used to produce a sample from the posterior}

\item{nIterations}{Number of Gibbs samples when method is Bayes, max. number of iterations 
when method is CML}
}
\value{
Depends on the estimation method
}
\description{
Fits the extended nominal response model by Bayesian sampling or CML
}
