% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_item_properties}
\alias{get_item_properties}
\title{Item properties in a project}
\usage{
get_item_properties(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A tibble with columns: item_property, type and values. Values shows 
unique values for your item properties, cut off if there are many.
}
\description{
Quickly glimpse the item properties defined in the project (if any).
}
