% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_persons}
\alias{get_persons}
\title{Persons in a project}
\usage{
get_persons(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame with columns person_id and columns for each person_property
}
\description{
Retrieve all persons/respondents that have been entered in the db
so far together with their properties
}
