% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_data.R
\name{get_resp_data}
\alias{get_resp_data}
\title{Functions for developers}
\usage{
get_resp_data(dataSrc, qtpredicate = NULL, extra_columns = NULL,
  summarised = FALSE, env = NULL, protect_x = TRUE,
  retain_person_id = TRUE, merge_within_person = FALSE,
  parms_check = NULL)
}
\arguments{
\item{dataSrc}{data.frame, numeric matrix, dexter database or `dx_resp_data` object}

\item{qtpredicate}{quoted predicate}

\item{extra_columns}{to be returned in addition to person_id, booklet_id, item_score, item_id}

\item{summarised}{if TRUE, no item scores are returned, just sumscores}

\item{env}{environment for evaluation of qtpredicate, defaults to caller environment}

\item{protect_x}{best set TRUE (default)}

\item{retain_person_id}{whether to retain the original person_id levels or just use arbitrary integers}

\item{merge_within_person}{merge different booklets for the same person together}

\item{parms_check}{data.frame of item_id, item_score to check for coverage of data}
}
\value{
list of type `dx_resp_data` containing:
when summarised=F
  x: tibble(person_id, booklet_id, item_id, item_score, booklet_score <, extra_columns>)
when summarised=T
  x: tibble(person_id, booklet_id, booklet_score <, extra_columns>)

design: tibble(booklet_id, item_id), sorted
}
\description{
Regular users are advised not to use these functions.
as incorrect use can easily crash your R-session or lead to unexpected results.
}
\details{
These functions are meant for people wishing to develop their own models based
on the data management structure of dexter. Very little input checking is performed,
the benefit is some extra speed over using regular dexter functions.

Not all aspects of the interface are completely stable yet
}
