% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_rules}
\alias{get_rules}
\title{Get scoring rules}
\usage{
get_rules(db)
}
\arguments{
\item{db}{a connection to a Dexter database}
}
\value{
data.frame of scoring rules containing columns: item_id, response, item_score
}
\description{
Retrieve the scoring rules currently present in the dexter project db
}
