% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_testscores}
\alias{get_testscores}
\title{Provide test scores}
\usage{
get_testscores(dataSrc, predicate = NULL)
}
\arguments{
\item{dataSrc}{Data source: a dexter project db handle or a data.frame}

\item{predicate}{An optional expression to filter data, if NULL all data is used}
}
\value{
A data frame showing person_id' s and their test scores
}
\description{
Supplies the weighted sum of item scores for each person selected.
}
