% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{get_variables}
\alias{get_variables}
\title{Variables that are defined in the project}
\usage{
get_variables(db)
}
\arguments{
\item{db}{a dexter project database}
}
\value{
a data.frame with name and type of the variables defined in your dexter project
}
\description{
Inspect the variables defined in your dexter project and their datatypes
}
\details{
The variables in Dexter consist of the item properties and person covariates you specified
and a number of reserved variables that are automatically defined like \code{response} and \code{booklet_id}.

Variables in Dexter are most useful when used in predicate expressions. A number of functions can 
take a dataSrc argument and an optional predicate. Predicates are 
a concise and flexible way to filter data for the different psychometric functions in Dexter.

The variables can also be used to retrieve data in \code{\link{get_responses}}
}
