% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{iModels}
\alias{iModels}
\title{Interactive model display}
\usage{
iModels(db, booklet)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{create_new_project}
or \code{open_project}}

\item{booklet}{ID of the booklet that will be shown}
}
\description{
Opens up a shiny application with item statistics and interactive
plots for the Rasch and Interaction models
}
