% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{iTIA}
\alias{iTIA}
\title{Interactive test-item analysis}
\usage{
iTIA(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{create_new_project}
or \code{open_project}}
}
\value{
An object that represents the application. Printing the object or passing it to \code{shiny::runApp} will run the app.
}
\description{
Open a shiny application for interactive item-test
analysis on the database
}
