% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{open_project}
\alias{open_project}
\title{Open an existing project}
\usage{
open_project(db_name = "dexter.db")
}
\arguments{
\item{db_name}{The name of the data base to be opened.}
}
\value{
A handle to the data base.
}
\description{
Opens a data base created by function \code{start_new_project}
}
\author{
Ivailo Partchev
}

