% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\docType{data}
\name{PISA_item_class}
\alias{PISA_item_class}
\title{Item properties in the PISA 2012 example}
\format{A data set with 109 rows and 6 columns.}
\description{
A data set of item properties in the PISA 2012 example (see the
help screen for function add_booklet)
}
\keyword{datasets}
