% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.R
\name{plot.p2pass}
\alias{plot.p2pass}
\title{A plot method for probability_to_pass}
\usage{
\method{plot}{p2pass}(x, ..., booklet_id = NULL, what = c("both",
  "equating", "sens/spec"))
}
\arguments{
\item{x}{An object produced by function \code{\link{probability_to_pass}}}

\item{...}{Any additional plotting parameters.}

\item{booklet_id}{vector of booklet_id's to plot, if NULL all booklets are plotted}

\item{what}{information to plot, 'equating', 'sens/spec' or 'both'}
}
\description{
Plot equating information from probability_to_pass
}
