% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{plot.prms}
\alias{plot.prms}
\title{Plot for the extended nominal Response model}
\usage{
\method{plot}{prms}(x, item_id = NULL, nbins = 5, ci = 0.95, ...)
}
\arguments{
\item{x}{object produced by fit_enorm}

\item{item_id}{which item to plot, if NULL, one plot for each item is made}

\item{nbins}{number of ability groups}

\item{ci}{confidence interval for the error bars, between 0 and 1. 0 means no error bars.
Default = 0.95 for a 95\% confidence interval}

\item{...}{further arguments to plot}
}
\description{
The plot shows 'fit' by comparing the expected score based on the model (grey line)
with the average scores based on the data (black line with dots) for groups of students
with similar estimated ability.
}
