% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3c.R
\name{plot3DC}
\alias{plot3DC}
\title{Show 3DC plots}
\usage{
plot3DC(db3dc, test_id = NULL, ...)
}
\arguments{
\item{db3dc}{3dc database handle}

\item{test_id}{optionally, a vector of test_id's. If omitted, plots for all tests will be shown}

\item{...}{further arguments to plot. Some of them have useful defaults}
}
\description{
Show 3DC plots as used in 3DC standard setting application
}
