% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate_variables}
\alias{predicate_variables}
\title{Deprecated alias for get_variables}
\usage{
predicate_variables(db)
}
\arguments{
\item{db}{a dexter project database}
}
\value{
a data.frame with name and type of the variables defined in your dexter project
}
\description{
This function has been renamed to \code{\link{get_variables}}. 
\code{Predicate_variables} will be removed in a future version.
}
