% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{print.prms}
\alias{print.prms}
\title{A print method for ENORM parms}
\usage{
\method{print}{prms}(x, ...)
}
\arguments{
\item{x}{An object produced by function \code{fit_enorm}}

\item{...}{Any other parameters to the print method}
}
\description{
A print method for ENORM parms
}
