% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{r_score_IM}
\alias{r_score_IM}
\title{Simulation from the interaction model}
\usage{
r_score_IM(m, scores)
}
\arguments{
\item{m}{an object produced by function \code{fit_inter}}

\item{scores}{vector of test scores}
}
\value{
a matrix with item scores, one column per item and one row per test score. Row order
equal to scores
}
\description{
Simulate item scores conditional on test scores using the interaction model
}
