% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{data}
\name{ratedData}
\alias{ratedData}
\title{Rated data}
\format{A data set with 75 rows and 15 columns.}
\description{
A data set with rated data. A number of student performances are rated twice on several
aspects by independent judges. The ratings are binary and have been summed following
the theory discussed by Maris and Bechger (2006, Handbook of Statistics). Data are a
small subset of data collected on the State Exam Dutch as a second language for Speaking.
}
\keyword{datasets}
