% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{data}
\name{ratedDataRules}
\alias{ratedDataRules}
\title{Scoring rules for the rated data}
\format{A data set with 42 rows and 3 columns (item_id, response, item_score).}
\description{
A set of (trivial) scoring rules for the rated data set
}
\keyword{datasets}
