% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_booklets}
\alias{show_booklets}
\title{List booklets in a project}
\usage{
show_booklets(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame showing the internal booklet number, the booklet name,
the number persons and the number of items
}
\description{
Show a list of the test forms (booklets) that have been entered in the db
so far
}
\author{
Ivailo Partchev
}

