% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_design}
\alias{show_design}
\title{Show the test design}
\usage{
show_design(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame (in string format) showing the booklet design: rows are items,
columns are booklets, and the numbers in the cells show the position of the item in the
booklet (blank if the booklet does not include the item)
}
\description{
Show a list of all items that have been entered in the db
so far by booklet and position in the booklet
}
