% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_items}
\alias{show_items}
\title{List items in a project}
\usage{
show_items(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame with items and item properties
}
\description{
Show all items that have been entered in the db
so far together with the item properties
}
