% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_person_properties}
\alias{show_person_properties}
\title{List person properties}
\usage{
show_person_properties(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data.frame of person properties, values and counts
}
\description{
Show all person properties defined in the project (if any)
}
