% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_persons}
\alias{show_persons}
\title{Show persons in a project}
\usage{
show_persons(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame with persons and covariates
}
\description{
Show all persons that have been entered in the db
so far together with covariates
}
