% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_rules}
\alias{show_rules}
\title{Show scoring rules}
\usage{
show_rules(db)
}
\arguments{
\item{db}{handle to a Dexter project database}
}
\value{
data.frame of scoring rules
}
\description{
Show the scoring rules currently present in the dexter project db
}
