% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_analysis}
\alias{univariate_analysis}
\title{Utility for doing univariate analysis}
\usage{
univariate_analysis(input_data, continuous_cols, continuous_threshold = 5)
}
\arguments{
\item{input_data}{Any object that can be converted into a data.table, will be what the univariate analysis is conducted on}

\item{continuous_cols}{A character vector containing the names of all columns that are considered continuous,
if not provided this will be calculated}

\item{continuous_threshold}{Integer, representing the number of unique values a column must have in order to be
considered continuous.  This will default to 5 and will only be used if continuous_cols was not provided.}
}
\value{
A list of summaries and plots.  The summaries are split into continuous_summaries for the summary statistics of
    continuous columns and categorical_summaries the categorical columns. The plots are divided into:
    \itemize{
      \item categorical_bar_charts - bar charts for categorical columns
      \item continuous_box_plots - box plots for continuous variables
      \item continuous_histograms - histograms for continuous variables
    }
}
\description{
\code{univariate_analysis} returns a list of summaries and plots for univariate analysis.
}
\details{
Used to help ease the process of doing univariate analysis.
For help with plotting of the results @seealso \code{\link{plot_univariates}}
}
\examples{
\dontrun{results = univariate_analysis(mtcars)}
\dontrun{results = univariate_analysis(mtcars, c('cyl', 'am', 'gear', 'carb'))}
\dontrun{results = univariate_analysis(mtcars, continuous_threshold=3)}
}
