% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{data}
\name{verbAggrData}
\alias{verbAggrData}
\title{Verbal aggression data}
\format{A data set with 316 rows and 25 columns.}
\description{
A data set of self-reported verbal behaviour in different frustrating
situations (Vansteelandt, 2000)
}
\keyword{datasets}

