% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{data}
\name{verbAggrRules}
\alias{verbAggrRules}
\title{Scoring rules for the verbal aggression data}
\format{A data set with 72 rows and 3 columns (item, response, score).}
\description{
A set of (trivial) scoring rules for the verbal 
aggression data set
}
\keyword{datasets}

