% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{as.data.frame.prms}
\alias{as.data.frame.prms}
\title{Coerce parameters object to a data.frame of parameters}
\usage{
\method{as.data.frame}{prms}(x, ...)
}
\arguments{
\item{x}{An object produced by function \code{fit_enorm}}

\item{...}{Any other parameters to the as.data.frame}
}
\description{
Coerce parameters object to a data.frame of parameters
}
