% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{close_project}
\alias{close_project}
\title{Close a project}
\usage{
close_project(db)
}
\arguments{
\item{db}{a Dexter project db handle}
}
\description{
This is just an alias for \code{DBI::dbDisconnect(db)}, included for completeness
}
