% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\docType{package}
\name{dexter-package}
\alias{dexter}
\alias{dexter-package}
\title{Dexter: data analyses for educational and psychological tests.}
\description{
Dexter provides a comprehensive solution for managing and analyzing educational test data.
}
\details{
The main features are:

\itemize{
\item project databases providing a structure for storing data about persons, items, responses and booklets.
\item methods to assess data quality using Classical test theory and plots.
\item CML calibration of the extended nominal response model and interaction model.
}

To learn more about dexter, start with the vignettes: `browseVignettes(package="dexter")`
}
\seealso{
Useful links:
\itemize{
  \item \url{http://dexterities.netlify.com}
  \item Report bugs at \url{https://github.com/jessekps/dexter/issues}
}

}
