% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{distractor_plot}
\alias{distractor_plot}
\title{Distractor plot}
\usage{
distractor_plot(dataSrc, item, predicate = NULL, nc = 1, nr = 1, ...)
}
\arguments{
\item{dataSrc}{Data source: a dexter project db handle or a data.frame with columns: person_id, item_id, response, item_score}

\item{item}{The ID of the item to plot. A separate plot will be produced
for each booklet that contains the item, or an error message if the item ID
is not known. Each plot contains a non-parametric regression of each possible
response on the total score.}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{nc}{An integer between 1 and 3. Number of columns when putting mutiple plots
on the same page. Default is 1. May be ignored or adjusted if it does not make sense.}

\item{nr}{An integer between 1 and 3. Number of rows when putting mutiple plots
on the same page. Default is 1. May be ignored or adjusted if it does not make sense.}

\item{...}{further arguments to plot.}
}
\description{
Produce a diagnostic distractor plot for an item
}
\details{
Customisation of title and subtitle can be done by using the arguments main and sub. 
These arguments can contain references to the variables item_id, booklet_id, item_position 
(only if dataSrc is a dexter db), pvalue, rit and rir. References are made by prefixing 
these variables with a dollar sign. Variable names can optionally be postfixed 
with a sprintf string, e.g.:
 
\code{distractor_plot(
db, main='item: $item_id', 
sub='Item-rest corr.: $rir:.2f')}
}
