% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{fit_domains}
\alias{fit_domains}
\title{Estimate the Rasch and the Interaction model per domain}
\usage{
fit_domains(dataSrc, item_property, predicate = NULL)
}
\arguments{
\item{dataSrc}{Data source: a dexter project db handle or a data.frame}

\item{item_property}{The item property defining the 
domains (subtests)}

\item{predicate}{An optional expression to subset data, if NULL all data is used}
}
\value{
An object of class \code{imp} holding results
for the Rasch model and the interaction model.
}
\description{
Estimate the parameters of the Rasch model and the Interaction model
}
\details{
Unlike the Rasch model, the interaction model cannot be computed
concurrently for a whole design of test forms. This function fits the
Rasch model and the interaction model on a complete rectangular array of
responses, with comparison between the two models playing an important role.
The rectangular array is typically one booklet but can also consist of 
the intersection (common items) of two or more booklets. If the intersection is empty 
(no common items for all persons), the function will exit with an error message.
}
\examples{
\dontrun{
db = start_new_project(verbAggrRules, "verbAggression.db")
add_booklet(db, verbAggrData, "agg")
add_item_properties(db, verbAggrProperties)
mSit = fit_domains(db, item_property= "situation")
plot(mSit)
}

}
\seealso{
\code{\link{plot.rim}}, \code{\link{fit_inter}}, \code{\link{add_item_properties}}
}
