% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{fit_enorm}
\alias{fit_enorm}
\title{Fit the extended nominal response model}
\usage{
fit_enorm(dataSrc, predicate = NULL, fixed_params = NULL,
  method = c("CML", "Bayes"), nIterations = 500,
  link_persons = FALSE)
}
\arguments{
\item{dataSrc}{Data source: a connection to a dexter database or a data.frame with columns: person_id, item_id, item_score}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{fixed_params}{Optionally, a prms object from a previous analysis or 
a data.frame with columns: item_id, item_score (omitting 0 score category) and beta. To facilitate the user in entering parameter values, we assume the parameterisation used by OPLM; in short, beta's are thresholds between categories. At this moment, it is not possible to fix some but not all categories of an item.}

\item{method}{If CML, the estimation method will be Conditional Maximum Likelihood;
otherwise, a Gibbs sampler will be used to produce a sample from the posterior}

\item{nIterations}{Number of Gibbs samples when estimation method is Bayes. The maximum 
number of iterations when using CML.}

\item{link_persons}{whether to merge different booklets administered to the same person, enabling linking over persons instead of just booklets.}
}
\value{
An object of type \code{prms}. The prms object can be cast to a data.frame of item parameters 
using function `coef` or used directly as input for other Dexter functions.
}
\description{
Fits an Extended NOminal Response Model (ENORM) using conditional maximum likelihood (CML)
or a Gibbs sampler for Bayesian estimation.
}
\references{
Maris, G., Bechger, T.M. and San-Martin, E. (2015) A Gibbs sampler for the (extended) marginal Rasch model. 
Psychometrika. 2015; 80(4): 859–879.
}
\seealso{
functions that accept a prms object as input: \code{\link{ability}}, \code{\link{plausible_values}}, 
\code{\link{plot.prms}}
}
