% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_booklets}
\alias{get_booklets}
\title{Booklets entered in a project}
\usage{
get_booklets(db)
}
\arguments{
\item{db}{a connection to a dexter database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame with columns: booklet_id, n_persons and n_items.
}
\description{
Retrieve information about the booklets entered in the db so far
}
