% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_item_properties}
\alias{get_item_properties}
\title{Item properties in a project}
\usage{
get_item_properties(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data.frame with one row per unique value of each item property and a count of the nbr of items it applies to.
The columns are: item_property, value, N
}
\description{
Retrieve information about the item properties defined in the project (if any).
This will return a data.frame with one row per unique value of each item property.
}
