% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_items}
\alias{get_items}
\title{Items in a project}
\usage{
get_items(db)
}
\arguments{
\item{db}{a connection to a dexter database, e.g. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data frame with column item_id and a column for each item property
}
\description{
Retrieve all items that have been entered in the db
so far together with the item properties
}
