% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_person_properties}
\alias{get_person_properties}
\title{Person properties in a project}
\usage{
get_person_properties(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A tibble with columns: person_property, type and values. Values shows unique values for your person properties, cut off if there are many.
}
\description{
Quickly glimpse the person properties defined in the project (if any).
}
