% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{get_testscores}
\alias{get_testscores}
\title{Provide test scores}
\usage{
get_testscores(dataSrc, predicate = NULL)
}
\arguments{
\item{dataSrc}{Data source: a connection to a dexter database or a data.frame with columns: person_id, booklet_id, item_score}

\item{predicate}{An optional expression to filter data, if NULL all data is used}
}
\value{
A tibble with columns person_id, item_id, booklet_score
}
\description{
Supplies the weighted sum of item scores for each person selected.
}
