% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{information}
\alias{information}
\alias{expected_score}
\alias{r_score}
\alias{p_score}
\title{Functions of theta}
\usage{
information(
  parms,
  items = NULL,
  booklet_id = NULL,
  parms_draw = c("average", "sample")
)

expected_score(
  parms,
  items = NULL,
  booklet_id = NULL,
  parms_draw = c("average", "sample")
)

r_score(
  parms,
  items = NULL,
  booklet_id = NULL,
  parms_draw = c("average", "sample")
)

p_score(
  parms,
  items = NULL,
  booklet_id = NULL,
  parms_draw = c("average", "sample")
)
}
\arguments{
\item{parms}{object produced by \code{\link{fit_enorm}} or a data.frame with columns item_id, item_score and, 
depending on parametrization, a column named either beta/delta, eta or b}

\item{items}{vector of one or more item_id's. If NULL and booklet_id is also NULL, all items in parms are used}

\item{booklet_id}{id of a single booklet (e.g. the test information function), if items is not NULL this is ignored}

\item{parms_draw}{when the item parameters are estimated with method "Bayes" (see: \code{\link{fit_enorm}}), 
parms_draw specifies whether to use a sample (a different item parameter draw for each output column) or the posterior mean
of the item draws. Alternatively, it can be an integer specifying a specific draw. It is ignored when parms is not estimated Bayesianly.}
}
\value{
Each function returns a new function which accepts a vector of theta's. These return the following values: 
\describe{
\item{information}{an equal length vector with the information estimate at each value of theta.}
\item{expected_score}{an equal length vector with the expected score at each value of theta}
\item{r_score}{a matrix with length(theta) rows and one column for each item containing simulated scores based on theta. 
To obtain test scores, use rowSums on this matrix}
\item{p_score}{a matrix with length(theta) rows and one column for each possible sumscore containing the probability of 
the score given theta}
}
}
\description{
returns information function, expected score function, score simulation function, or score distribution 
for a single item, an arbitrary group of items or all items
}
\examples{

\dontshow{ RcppArmadillo::armadillo_throttle_cores(1)}

db = start_new_project(verbAggrRules,':memory:')
add_booklet(db,verbAggrData, "agg")
p = fit_enorm(db)

# plot information function for single item

ifun = information(p, "S1DoScold")

plot(ifun,from=-4,to=4)

# compare test information function to the population ability distribution

ifun = information(p, booklet="agg")

pv = plausible_values(db,p)

op = par(no.readonly=TRUE)
par(mar = c(5,4,2,4))

plot(ifun,from=-4,to=4, xlab='theta', ylab='test information')

par(new=TRUE)

plot(density(pv$PV1), col='green', axes=FALSE, xlab=NA, ylab=NA, main=NA)
axis(side=4)
mtext(side = 4, line = 2.5, 'population density (green)')

par(op)
close_project(db)

\dontshow{ RcppArmadillo::armadillo_reset_cores()}

}
