% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{iTIA}
\alias{iTIA}
\title{Interactive test-item analysis}
\usage{
iTIA(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{create_new_project}
or \code{open_project}}
}
\description{
Opens up a shiny application to do interactive item-test
analysis on the database
}
