% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{latent_cor}
\alias{latent_cor}
\title{Latent correlations}
\usage{
latent_cor(
  dataSrc,
  item_property,
  predicate = NULL,
  nDraws = 500,
  use = "complete.obs"
)
}
\arguments{
\item{dataSrc}{A connection to a dexter database or a data.frame with columns: person_id, item_id, item_score and 
the item_property}

\item{item_property}{The name of the item property used to define the domains. If \code{dataSrc} is a dexter db then the
item_property must match a known item property. If datasrc is a data.frame, item_property must be equal to
 one of its column names.}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{nDraws}{Number of draws for plausible values}

\item{use}{Only complete.obs at this time. Respondents who don't have a score for one or more scales are removed.}
}
\value{
List containing a correlation matrix and corresponding standard deviations
}
\description{
Estimates correlations between latent traits. Use an item_property to distinguish the different scales. 
This function uses plausible values so results may differ slightly between calls. 
Note: this is a new and slightly experimental function and therefore still a bit slow. It will probably 
become faster in future versions of dexter.
}
