% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enorm.R
\name{latent_cor}
\alias{latent_cor}
\title{Estimates latent correlations}
\usage{
latent_cor(
  dataSrc,
  item_property,
  predicate = NULL,
  quick = FALSE,
  nIterations = 100,
  use = "complete.obs"
)
}
\arguments{
\item{dataSrc}{a connection to a dexter database or a data.frame with columns: person_id, item_id, item_score}

\item{item_property}{An item property to distinguish the different scales.}

\item{predicate}{An optional expression to subset data, if NULL all data is used}

\item{quick}{If true, a estimate based on attenuated correlations between plausible values is produced.}

\item{nIterations}{Number of iterations for plausible values}

\item{use}{only complete.obs at this time. Respondents who don't have a score for one or more scales are removed.}
}
\value{
Correlation matrix and corresponding standard deviations
}
\description{
Estimates latent correlations
}
