% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{open_project}
\alias{open_project}
\title{Open an existing project}
\usage{
open_project(db_name = "dexter.db")
}
\arguments{
\item{db_name}{The name of the database to be opened.}
}
\value{
a database connection object
}
\description{
Opens a database created by function \code{start_new_project}
}
