% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{plot.DIF_stats}
\alias{plot.DIF_stats}
\title{plot method for pairwise DIF statistics}
\usage{
\method{plot}{DIF_stats}(
  x,
  items = NULL,
  itemsX = items,
  itemsY = items,
  cluster = FALSE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{x}{object produced by \code{\link{DIF}}}

\item{items}{character vector of item id's for a subset of the plot. Useful if you have many items. 
If NULL all items are plotted.}

\item{itemsX}{character vector of item id's for the X axis}

\item{itemsY}{character vector of item id's for the Y axis}

\item{cluster}{arrange items by similarity.}

\item{alpha}{significance level used to color the plot (two sided)}

\item{...}{further arguments to plot}
}
\description{
plot method for pairwise DIF statistics
}
\details{
Plotting produces an image of the matrix of pairwise DIF statistics. 
The statistics are standard normal deviates and colored to distinguish significant from non-significant values.
If there is no DIF, a proportion alpha off the cells will be colored significant by chance alone.
}
