% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.R
\name{plot.p2pass}
\alias{plot.p2pass}
\title{A plot method for probability_to_pass}
\usage{
\method{plot}{p2pass}(x, what = c("all", "equating", "sens/spec", "roc"),
  booklet_id = NULL, ...)
}
\arguments{
\item{x}{An object produced by function \code{\link{probability_to_pass}}}

\item{what}{information to plot, 'equating', 'sens/spec', 'roc, or 'all'}

\item{booklet_id}{vector of booklet_id's to plot, if NULL all booklets are plotted}

\item{...}{Any additional plotting parameters; e.g., cex = 0.7.}
}
\description{
Plot equating information from probability_to_pass
}
