% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{predicate_variables}
\alias{predicate_variables}
\title{Variables that can be used in expressions}
\usage{
predicate_variables(db)
}
\arguments{
\item{db}{a dexter project database}
}
\value{
a data.frame with name and type of the variables defined in your dexter project
}
\description{
Show the variables defined in your dexter project and their datatypes
}
\details{
A number of functions can take a dataSrc arguments and an optional expression. Expressions are a concise and
flexible way to filter data for the different psychometric functions in Dexter.
The variables that you can use in expressions consist of item properties and person covariates you defined
and a number of reserved variables that are automatically defined like \code{response} and \code{booklet_id}.
}
