% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_model.R
\name{print.rim}
\alias{print.rim}
\title{A print method for the interaction model}
\usage{
\method{print}{rim}(x, ...)
}
\arguments{
\item{x}{An object produced by function \code{fit_inter}}

\item{...}{Further arguments to print}
}
\description{
Print the available items for plots of the Rasch and the interaction models
}
