% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{profile_plot}
\alias{profile_plot}
\title{Profile plot}
\usage{
profile_plot(db, booklets, domains, grpVar, group = NULL, model = "IM")
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}

\item{booklets}{A vector with the numbers of the test forms (booklets)
to be included in the analysis. See Details.}

\item{domains}{The item property defining the domains}

\item{grpVar}{Variable name for subsetting}

\item{group}{Value(s) for subsetting; if NULL, all distinct values
will be used.}

\item{model}{"IM" (default) or "RM"}
}
\value{
Nothing interesting
}
\description{
Profile plot
}
\author{
Timo Bechger, Gunter Maris, Ivailo Partchev
}

