% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{show_item_properties}
\alias{show_item_properties}
\title{List item properties}
\usage{
show_item_properties(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A data.frame of item properties, values and counts
}
\description{
Show a list of the item properties defined in the project (if any)
}
