% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexter.R
\name{tia_tables}
\alias{tia_tables}
\title{Simple Test-Item Analysis}
\usage{
tia_tables(db)
}
\arguments{
\item{db}{A handle to the database, i.e. the output of \code{start_new_project}
or \code{open_project}}
}
\value{
A list of two data frames, itemStats and testStats.
}
\description{
Produce tables with simple test-item analysis statistics
}
\details{
When an item is contained in more than one booklet, the 
item statistics are averaged across booklets with the number of 
persons in each booklets serving as weights.
}
\author{
Ivailo Partchev
}

